+++ src/utils.js
@@ -51,6 +51,14 @@ export function prepareDailyResponse(response) {
   })
 }
 
+export function prepareHydrationResponse(response) {
+  return response.map(el => {
+    el.date = new Date(el.date).toISOString()
+    el.waterConsumed = Number(el.waterConsumed).toFixed(3)
+    return el
+  })
+}
+
 function getFormattedDate(date) {
   const day = ('0' + date.getDate()).slice(-2)
   const month = ('0' + (date.getMonth() + 1)).slice(-2)